Split Apps components
